Move mapping functions into common util file
